# wkt: improvements