… planar/makevalid/setdiff/polygoncleaner_test.go
… planar/triangulate/constraineddelaunay/triangulator_test.go
… planar/triangulate/delaunay_test.go
… planar/triangulate/gdey/quadedge/main_test.go
… planar/triangulate/quadedge/quadedgesubdivision_test.go
… windingorder/windingorder_test.go