## performance
This pr has a pretty good performance improvement over the previous encoder, both in cpu time and allocations. The most important function in all of this is `formatFloat` for an even greater performance improvement it could be written in assembly or maybe use the 'e' formatter.

A simple change that came to mind is also:
```go
type Encoder struct {
   w io.Writer
   buf []byte // set to make([]byte, 0, 16)
}


func (enc *Encoder) formatFloat(f float64) error {
	// for reference, the min value of EPSG:3857 -20026376.390 -> 13 characters
	buf := strconv.AppendFloat(enc.buf, f, 'f', 3, 64)
	i := len(buf) - 1;
	for ; i >= 0; i-- {
		if buf[i] != '0' {
			break
		}
	}
	if buf[i] == '.' {
		i--
	}
	buf = buf[:i+1]
	_, err := enc.w.Write(buf)
	enc.buf = enc.buf[:0]
	return err
}
```

## benchmarks
I'm using the following command for running benchmarks:
```
go test -v -run=xxx -bench=. --benchtime=5s -benchmem
```

### before
```
goos: darwin
goarch: amd64
pkg: github.com/go-spatial/geom/encoding/wkt
BenchmarkEncodeSin100-4    	   88194	     67946 ns/op	   22624 B/op	     614 allocs/op
BenchmarkEncodeSin1000-4   	    8854	    681131 ns/op	  214720 B/op	    6017 allocs/op
BenchmarkEncodeTile-4      	       6	 852446852 ns/op	597514269 B/op	 7658125 allocs/op
PASS
ok  	github.com/go-spatial/geom/encoding/wkt	18.828s
```
### after
```
goos: darwin
goarch: amd64
pkg: github.com/go-spatial/geom/encoding/wkt
BenchmarkEncodeSin100-4    	  108333	     56824 ns/op	    9817 B/op	     411 allocs/op
BenchmarkEncodeSin1000-4   	   10000	    549257 ns/op	   87777 B/op	    4014 allocs/op
BenchmarkEncodeTile-4      	       8	 639061782 ns/op	123882434 B/op	 5041352 allocs/op
PASS
ok  	github.com/go-spatial/geom/encoding/wkt	18.062s
```

### latest
```goos: darwin
goarch: amd64
pkg: github.com/go-spatial/geom/encoding/wkt
BenchmarkEncodeSin100-4           	  128605	     46471 ns/op	    6000 B/op	       6 allocs/op
BenchmarkEncodeSin1000-4          	   12789	    471229 ns/op	   48496 B/op	       8 allocs/op
BenchmarkEncodeTile-4             	      10	 513830525 ns/op	67322720 B/op	     112 allocs/op
BenchmarkEncodeTilePrealloc-4     	      10	 502564612 ns/op	33556064 B/op	      99 allocs/op
BenchmarkEncodeTileNoprealloc-4   	      10	 528629372 ns/op	104913264 B/op	     120 allocs/op
PASS
ok  	github.com/go-spatial/geom/encoding/wkt	34.258s
```

### latest with 'e' parameter
```
goos: darwin
goarch: amd64
pkg: github.com/go-spatial/geom/encoding/wkt
BenchmarkEncodeSin100-4           	  302805	     19736 ns/op	    6000 B/op	       6 allocs/op
BenchmarkEncodeSin1000-4          	   30608	    196121 ns/op	   89456 B/op	       9 allocs/op
BenchmarkEncodeTile-4             	      24	 219000498 ns/op	67322720 B/op	     112 allocs/op
BenchmarkEncodeTilePrealloc-4     	      28	 207262018 ns/op	33556064 B/op	      99 allocs/op
BenchmarkEncodeTileNoprealloc-4   	      26	 225557254 ns/op	100664923 B/op	     120 allocs/op
PASS
ok  	github.com/go-spatial/geom/encoding/wkt	31.788s
```

## usability
My implementation is a bit longer because it has to deal with a choice of removing or keeping empty geometries. It also returns a proper error message when lines and geometries with not enough valid points are sent for a polygon or linestring.

I just noticed there is also a bug, where empty points can be sent on a line string, this should not be allowed.

## extra
I also took the liberty of changing all "Verticies" to "Vertices", I can remove this commit.


## todo
- [x] preallocate buffer in `Encoder` struct for `formatFloat`
- [x] refactor package so tests pass
- [x] properly handle empty points in linestrings and polygons