@ear7h 2019-09-24
@gdey yep and ci should pass. I removed the commit which implements build tags for tile test files. I will implement the solution we discussed in a different pr (#72)